refactored commands
