+++ .env.sample
@@ -1,1 +1,1 @@
-DESKTOP_ASSISTANT_SMTP_PWD=YOUR_EMAIL_PASSWORD
+SMTP_PASSWORD=YOUR_EMAIL_PASSWORD
