--- backups/1.0.0/README.md
@@ -1,45 +1,0 @@
-# 🗂️ Backups Module - Version 1.0.0
-
-This module contains the backup implementation of the Desktop Assistant project.
-
-![Version](https://img.shields.io/badge/Version-1.0.0-green.svg)
-
-## 📁 Version Structure
-
-The `backups/1.0.0` directory contains the following files and directories:
-
-- `README.md`: This file, containing details about the module.
-- `src/`: Source code directory for the backup implementation.
-- `setupProject.sh`: Script to set up the project environment.
-- `requirements.txt`: File containing the required Python packages for the project.
-
-## 🛠️ Setup
-
-To set up the project, follow these steps:
-
-1. **Clone the repository**:
-    ```bash
-    git clone https://github.com/vihar-s1/Desktop-Assistant
-    ```
-
-2. **Navigate to the backup directory**:
-    ```bash
-    cd Desktop-Assistant/backups/1.0.0
-    ```
-
-3. **Run the setup script**:
-    ```bash
-    bash setupProject.sh
-    ```
-
-4. **Activate the virtual environment**:
-    ```bash
-    source .venv/bin/activate
-    ```
-
-## ▶️ How to Run the Project
-
-After setting up the project, you can run the assistant using the following command:
-
-```bash
-python3 assistant.py
