--- backups/1.0.0/requirements.txt
@@ -1,34 +1,0 @@
-#
-# This file is autogenerated by pip-compile with Python 3.11
-# by the following command:
-#
-#    pip-compile --no-annotate --output-file=requirements_temp.txt requirements.txt
-#
-appopener~=1.6
-beautifulsoup4~=4.12.2
-build~=1.0.3
-certifi~=2024.7.4
-charset-normalizer~=3.2.0
-click~=8.1.7
-colorama~=0.4.6
-comtypes~=1.2.0
-googlesearch-python~=1.2.3
-idna~=3.4
-packaging~=23.1
-pillow~=11.0.0
-pip-tools~=7.3.0
-pipdeptree~=2.13.0
-pyaudio~=0.2.13
-pyautogui~=0.9.54
-pyproject-hooks~=1.0.0
-pyttsx3~=2.90
-requests~=2.32.0
-soupsieve~=2.5
-speechrecognition~=3.10.0
-urllib3~=2.2.2
-wheel~=0.41.2
-wikipedia~=1.4.0
-
-# The following packages are considered to be unsafe in a requirements file:
-# pip
-# setuptools
