+++ requirements.txt
@@ -1,51 +1,16 @@
-#
-# This file is autogenerated by pip-compile with Python 3.11
-# by the following command:
-#
-#    pip-compile --no-annotate --output-file=requirements_temp.txt requirements.txt
-#
-appopener==1.7
-beautifulsoup4==4.12.3
-build==1.0.3
-certifi==2024.7.4
-charset-normalizer==3.2.0
-click==8.1.8
-colorama==0.4.6
-comtypes==1.2.1
-feedparser==6.0.11
-googlesearch-python==1.2.5
-idna==3.10
-mouseinfo==0.1.3
-packaging==23.2
-pillow==11.0.0
-pip-tools==7.3.0
-pipdeptree==2.13.2
-psutil==6.1.1
-pyaudio==0.2.14
-pyautogui==0.9.54
-pycaw==20240210
-pygetwindow==0.0.9
-pymsgbox==1.0.9
-pyperclip==1.9.0
-pypiwin32==223
-pyproject-hooks==1.0.0
-pyrect==0.2.0
-pyscreeze==1.0.1
-python-dotenv==1.0.1
-pyttsx3==2.98
-pytweening==1.2.0
-pywin32==308
-regex==2024.11.6
-requests==2.32.3
-sgmllib3k==1.0.0
-soupsieve==2.6
-speechrecognition==3.10.4
-typing-extensions==4.12.2
-urllib3==2.2.3
-wheel==0.41.3
-wikipedia==1.4.0
-wmi==1.5.1
-
-# The following packages are considered to be unsafe in a requirements file:
-# pip
-# setuptools
+appopener~=1.6
+colorama~=0.4.6
+feedparser~=6.0.11
+googlesearch-python~=1.3.0
+pillow~=11.1.0
+pip-tools~=7.4.1
+pipdeptree~=2.25.0
+PyAudio~=0.2.14
+PyAutoGUI~=0.9.54
+pycaw~=20240210
+python-dotenv~=1.0.1
+pyttsx3~=2.98
+regex~=2024.11.6
+SpeechRecognition~=3.14.1
+wikipedia~=1.4.0
+WMI~=1.4.9
